fix: merge dual User models, fix timedelta import, move billing models (CA-001, CA-002, CA-018)
CA-002: Unified app/models/user.py now contains all auth + billing columns
(id, username, email, hashed_password, role, plan_id, stripe_customer_id,
is_active, created_at, updated_at) with relationships to billing models.
Eliminates SQLAlchemy mapper collision between the two competing User classes.

CA-001: Fixed timedelta NameError in Analysis.expires_at default lambda.
Import corrected in new app/models/billing.py.

CA-018: Moved Analysis, APIKey, UsageLog, BillingEvent to app/models/billing.py.
app/database/models.py replaced with safe empty stub. Both app/main.py and
tests/conftest.py updated to import billing so tables are created at startup.

MIGRATION REQUIRED ON PROD SERVER before deploying:
  ALTER TABLE users ADD COLUMN IF NOT EXISTS plan_id VARCHAR(50) DEFAULT 'free';
  ALTER TABLE users ADD COLUMN IF NOT EXISTS stripe_customer_id VARCHAR(255);
  ALTER TABLE users ADD COLUMN IF NOT EXISTS is_active BOOLEAN DEFAULT TRUE;
  ALTER TABLE users ADD COLUMN IF NOT EXISTS created_at TIMESTAMP DEFAULT NOW();
  ALTER TABLE users ADD COLUMN IF NOT EXISTS updated_at TIMESTAMP DEFAULT NOW();
