+++ app/models/user.py
@@ -1,16 +1,32 @@
-"""User model with role-based access."""
+"""User model — unified auth + billing columns."""
 
 import uuid
-from sqlalchemy import Column, String, DateTime, func
+from sqlalchemy import Column, String, Boolean, DateTime, func
+from sqlalchemy.orm import relationship
 from app.database.session import Base
 
 
 class User(Base):
     __tablename__ = "users"
 
+    # Auth columns (original auth model)
     id = Column(String, primary_key=True, default=lambda: str(uuid.uuid4()))
     username = Column(String, unique=True, nullable=False, index=True)
     email = Column(String, unique=True, nullable=False)
     hashed_password = Column(String, nullable=False)
     role = Column(String, default="procurement")  # admin | procurement | legal
+
+    # Billing columns (from database/models.py User — CA-002)
+    plan_id = Column(String(50), default="free")  # free | starter | growth | business
+    stripe_customer_id = Column(String(255), nullable=True, unique=True)
+    is_active = Column(Boolean, default=True)
+
+    # Timestamps
     created_at = Column(DateTime, server_default=func.now())
+    updated_at = Column(DateTime, server_default=func.now(), onupdate=func.now())
+
+    # Relationships to billing models (declared in app/models/billing.py)
+    api_keys = relationship("APIKey", back_populates="user", cascade="all, delete-orphan")
+    analyses = relationship("Analysis", back_populates="user", cascade="all, delete-orphan")
+    usage_logs = relationship("UsageLog", back_populates="user", cascade="all, delete-orphan")
+    billing_events = relationship("BillingEvent", back_populates="user", cascade="all, delete-orphan")
